+++ test/test_win.py
@@ -36,7 +36,7 @@ class BaseTestWin(object):
         self.assertEqual(base, self.WIN.Get_attr(MPI.WIN_BASE))
 
     def testMemory(self):
-        memory = self.WIN.memory
+        memory = self.WIN.tomemory()
         pointer = MPI.Get_address(memory)
         length = len(memory)
         base, size, dunit = self.WIN.attrs
@@ -129,7 +129,7 @@ class BaseTestWinAllocate(BaseTestWin):
 
     def setUp(self):
         self.WIN = MPI.Win.Allocate(10, 1, self.INFO, self.COMM)
-        self.memory = self.WIN.memory
+        self.memory = self.WIN.tomemory()
         memzero(self.memory)
 
     def tearDown(self):
@@ -141,14 +141,14 @@ class BaseTestWinAllocateShared(BaseTestWin):
 
     def setUp(self):
         self.WIN = MPI.Win.Allocate_shared(10, 1, self.INFO, self.COMM)
-        self.memory = self.WIN.memory
+        self.memory = self.WIN.tomemory()
         memzero(self.memory)
 
     def tearDown(self):
         self.WIN.Free()
 
     def testSharedQuery(self):
-        memory = self.WIN.memory
+        memory = self.WIN.tomemory()
         address = MPI.Get_address(memory)
         length = len(memory)
         memories = self.COMM.allgather((address, length))
@@ -180,7 +180,7 @@ class BaseTestWinCreateDynamic(BaseTestWin):
         self.assertEqual(size, 0)
 
     def testMemory(self):
-        memory = self.WIN.memory
+        memory = self.WIN.tomemory()
         base = MPI.Get_address(memory)
         size = len(memory)
         self.assertEqual(base, 0)